iio/adc/ad9361: New device attribute xo_correction (offset freq correction)
Fixed frequency oscillators typically have a frequency offset from is
nominal value. Its guaranteed to be within the accuracy or tolerance.
However it’s often desirable to measure the actual frequency and then
use the measured instead of the nominal rate. In the context of the AD9361
reference errors are multiplied by the RF and BB PLLs. In situations where
the DCXO can't be used or this option allows or frequency corrections.

The max offset that can be compensated is limited by the "adjustable-clock"
clock-accuracy.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>