iio: ad9523: Fix race condition between setup sequence and clock callbacks
When running the initial register setup sequence the ad9523 driver
registers the clock that is associated with a specific output directly
after the output has been configured.

This opens up a race condition window since the clock ops might be invoked
once the clock has been registered. And while there is mutual exclusion
between any two clocks do register access there is no mutual exclusion
between the setup sequence and the clock operations.

As a result it is possible that register access from a clock operation
callback and the setup sequence happen concurrently which causes transfer
buffer corruption and results in undefined behavior.

To avoid this make sure that the initial register setup sequence has been
completed before registering the clocks. This will ensure that concurrent
access can not occur.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>