iio: ad9528: Make ID register read 3 bytes
Currently the ad9528 does a 4 byte read when reading the chip ID. The last
byte is the chip revision which is not checked anyway though.

Since the ad9528_read() function returns a signed 32 bit value and uses
negative values for error codes having a 4 byte register read where the MSB
is one results in an incorrectly reported error.

So change the ID register read to a 3 byte read.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>