net: ethernet: xilinx: move init of internal phy before external
When the MAC type is set to XAXIENET_1G or XAXIENET_2_5G,
the device connects to the PHY, skipping the initialization
of the internal PHY (if that is needed).

We evaluate the condition of `lp->phy_node_int` being
non-NULL, since this is initialized for SGMII operation
mode, which assumes that the PHY is external.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>