iio:adxl372: Add FIFO and interrupts support
This patch adds support for the adxl372 FIFO. In order to accomplish this,
triggered buffers were used.

The number of FIFO samples which trigger the watermark interrupt can be
configured by using the buffer watermark, while the format depends on the
selected channels.The FIFO data along with the timestamp is pushed to the
IIO device's buffer.

This patch also modifies the data.d8 array size from 512 to 1024, as this
buffer is used to store the FIFO data; the FIFO size is 512 samples of 2
bytes each.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>