iio:adxl372: Add sampling frequency support
This patch adds the option for the user to select the sampling frequency.
Also, the user can read the available frequencies and read the currently
set frequency via the read_raw function. The frequency can be set via the
write_raw function.

When the frequency is set, the bandwidth is also checked and ensured that
it is constrained to at most half of the sampling frequency.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>